+++ tests/TestCase/Config/GlobalFunctionsTest.php
@@ -275,6 +275,20 @@ class GlobalFunctionsTest extends TestCase
         ];
         $this->assertEquals($expected, $result);
         
+        $result = optionDefaults(['class' => 'first-class second-class'], $options);
+        $expected = [
+            'value1' => 'val-1',
+            'class' => 'first-class second-class',
+        ];
+        $this->assertEquals($expected, $result);
+        
+        $result = optionDefaults(['class' => ['first-class', 'second-class']], $options);
+        $expected = [
+            'value1' => 'val-1',
+            'class' => 'first-class second-class',
+        ];
+        $this->assertEquals($expected, $result);
+        
         //This doesn't change the value
         $result = optionDefaults(['value1' => 'new-val-1'], $options);
         $expected = ['value1' => 'val-1'];
@@ -285,6 +299,14 @@ class GlobalFunctionsTest extends TestCase
         $expected = [
             'value1' => 'val-1',
             'class' => 'my-class',
+        ];
+        $this->assertEquals($expected, $result);
+        
+        //Backward compatibility with three arguments
+        $result = optionDefaults('class', 'first-class second-class', $options);
+        $expected = [
+            'value1' => 'val-1',
+            'class' => 'first-class second-class',
         ];
         $this->assertEquals($expected, $result);
     }